Add processing init methods and unified sizes.

+++ Codec23200PCM.h
@@ -26,9 +26,9 @@ class CCodec23200PCM : public IProcessor {
     CCodec23200PCM();
     virtual ~CCodec23200PCM();
 
-    virtual uint8_t input(const uint8_t* buffer, uint8_t length);
+    virtual uint8_t  input(const uint8_t* buffer, uint8_t length);
 
-    virtual uint8_t output(uint8_t* buffer);
+    virtual uint16_t output(uint8_t* buffer);
 
   private:
 };
