+++ PCMCodec23200.h
@@ -26,9 +26,9 @@ class CPCMCodec23200 : public IProcessor {
     CPCMCodec23200();
     virtual ~CPCMCodec23200();
 
-    virtual uint8_t input(const uint8_t* buffer, uint8_t length);
+    virtual uint8_t  input(const uint8_t* buffer, uint8_t length);
 
-    virtual uint8_t output(uint8_t* buffer);
+    virtual uint16_t output(uint8_t* buffer);
 
   private:
 };
